Struct rubble_templates::evaluator::SyntaxError

pub struct SyntaxError {
    pub relative_pos: Position,
    pub invocation_pos: Position,
    pub description: EvaluationError,
}

An error that can happen during evaluation with full info about where and what happened.

Contains additional fields that describe EvaluationError. This is a wrapper of EvaluationError that provides more info about the error so the user can see where the error is and how they can fix it.

Unlike EvaluationError, it can be created where a much broader context is available and when additional info can be supplied (eg. position of currently evaluated block).

Fields
relative_pos: Positioninvocation_pos: Positiondescription: EvaluationError
Implementations

impl SyntaxError

pub fn new(error: EvaluationError) -> SyntaxError

Creates new SyntaxError with given EvaluationError.

Preferred way to create SyntaxError when the position of error is unknown. When the relative position is known, please construct this using at_pos.

pub fn at_position(position: Position, error: EvaluationError) -> SyntaxError

Creates new SyntaxError with given EvaluationError at known relative position..