Trait rubble_templates::evaluator::Function

pub trait Function {
    fn evaluate(
        &self, 
        evaluator: &dyn Evaluator, 
        parameters: &[SyntaxNode], 
        context: &mut Context
    ) -> Result<String, SyntaxError>;
}

A function that can be used to add features to the template.

Any struct that implements this trait adds additional features that can additional function which will be available to anyone using given Evaluator that has it installed.

For example, you might implement custom function for date parsing etc etc.

During evaluation, an original Evaluator will be supplied to enable parameter evaluation. Parameter evaluation with a supplied Evaluator is optional and a given Function can evaluate them independently.

Required methods

fn evaluate(
&self,
evaluator: &dyn Evaluator,
parameters: &[SyntaxNode],
context: &mut Context
) -> Result<String, SyntaxError>

Loading content...