Module rubble_templates::std_fun::math

Provides standard math functions.

See std_fun or math_functions for more info.

Functions

as_numbers

divide_function
Divides values. If any of the parameters is not convertible to a number, then an error will be emitted with the invalid value.

math_functions
Provides a set of math functions.

minus_function
Subtracts values. If any of the parameters is not convertible to a number, then an error will be emitted with the invalid value.

modulo_function
Calculates the remainder (modulo). If any of the parameters is not convertible to a number, then an error will be emitted with the invalid value.

multiply_function
Multiplies values. If any of the parameters is not convertible to a number, then an error will be emitted with the invalid value.

plus_function
Adds (or concatenates) values. If any of the parameters is not convertible to a number, then the rest will be concatenated.

reduce_numbers